+++ src/novelystlib/novelyst_tk.py
@@ -280,6 +280,7 @@ class NovelystTk(MainTk):
         self.exportMenu.add_command(label=_('Brief synopsis (export only)'), command=lambda: self._export_document('_brf_synopsis', lock=False))
         self.exportMenu.add_command(label=_('Cross references (export only)'), command=lambda: self._export_document('_xref', lock=False))
         self.exportMenu.add_separator()
+        self.exportMenu.add_command(label=_('Obfuscated text for word count'), command=lambda: self._export_document('_wrimo', lock=False, show=False))
         self.exportMenu.add_command(label=_('Characters/locations/items data files'), command=lambda: self._export_document('_data', lock=False, show=False))
 
         # Tools